Don't autoload 'lib'; explicitly 'require' things instead.
The new Rails 3 idiom is to only autoload things in the 'app'
directory. Instead, we 'require' the necessary modules/classes
where they are needed.

The reason is that autoloaded files in development become
eagerloaded files in production, and you don't generally want
everything in 'lib' automatically loaded when the app starts up;
that's what 'config/initializers' is for.

This commit tidies up 'lib' and 'config/initializers' to fit this
division of labour more closely.